fix: Spawner

+++ src/classes/AnimeDesu.ts
@@ -1,14 +1,14 @@
-import { Collection, MessageEmbed, TextChannel } from "discord.js";
+import { MessageEmbed, TextChannel } from "discord.js";
 import OtakuDesu from "otakudesu-scrape";
 import type { Desu } from "../typings";
 import type AnimeClient from "./AnimeClient";
 const otakudesu = new OtakuDesu();
 
 async function generate() {
-    const animes = await otakudesu.home();
+    const animes = await otakudesu.ongoing();
     const randomIndex = Math.floor(Math.random() * animes.length);
     const anime = await otakudesu.anime(animes[randomIndex].parse_name);
-    const filterNames = anime.title.split(' ').filter(t => /^[\w.-]+$/gi.test(t));
+    const filterNames = anime.details.judul.split(' ').filter(t => /^[\w.-]+$/gi.test(t));
     return {
         parse: animes[randomIndex].parse_name,
         regex: new RegExp(`(${filterNames.join('|')})`, 'gi'),
@@ -18,7 +18,7 @@ async function generate() {
 
 class AnimeDesu {
     constructor(private client: AnimeClient) {}
-    public works: Collection<string, Desu> = new Collection();
+    public works: Map<string, Desu> = new Map();
 
     public getGuild(guildID: string) {
         return this.works.get(guildID);
@@ -28,11 +28,11 @@ class AnimeDesu {
         if (work) return false;
         const anime = await generate();
         function censorTitle(str: string) {
-            return str[3] + '*'.repeat(str.length - 2) + str.slice(-1);
+            return str[0] + '*'.repeat(str.length - 2) + str.slice(-1);
         }
         const embed = new MessageEmbed()
         .setColor('RANDOM')
-        .setTitle(censorTitle(anime.title))
+        .setTitle(censorTitle(anime.details.judul))
         .setImage(anime.image)
         .setDescription('Ayo tebak ini anime apa?')
         .setTimestamp();
@@ -40,14 +40,16 @@ class AnimeDesu {
         this.works.set(guildID, {
             regex: anime.regex,
             anime: {
-                title: anime.title,
+                title: anime.details.judul,
                 parse: anime.parse,
                 image: anime.image
             },
             m,
             timeout: setTimeout(() => {
+                if (!this.getGuild(guildID)) return;
+                this.remGuild(guildID);
                 m.delete();
-                channel.send('Kayaknya gada yang berhasil jawab deh <3\nJawabannya adalah: ' + anime.title);
+                channel.send('Kayaknya gada yang berhasil jawab deh <3\nJawabannya adalah: ' + anime.details.judul);
             }, 30 * 1000)
         });
         return true;
